refactor: Re-arrange unneccesary folders
