+++ src/ui/FolderInfo.tsx
